+++ pkg/policies/flowcontrol/api/base/base.go
@@ -1,4 +1,4 @@
-package common
+package base
 
 import (
 	"context"
