+++ pkg/policies/flowcontrol/api/base/provide.go
@@ -1,4 +1,4 @@
-package common
+package base
 
 import (
 	"fmt"
